Change error metrics in scripts, figures, & tables
R2 is now called the nash sutcliffe coeficient of efficiency, NSE,
note the math is the same

RMSE is removed since it isn't comparable across sites. replaced
by the mean coefficient of variation of the mean absolute error,
aka the mean cvmae, aka F. This is also the loss metric used in
fitting.